Rethink the entire notification scheduling/handling/sending process
The old process checked which notification to send at least three times, which caused a lot of issues. In this implementation, the notify_class is defined only once, which is then passed around. So far, this seems to have fixed effectively all of the issues stemming from the previous implementation. Also, this resulted in quite a bit of removed semi-duplicate code.

This release can also successfully send a notification: currently only to the owner of the bot (as in none of the previous checks/db pulls aren't implemented for 1.6 yet), but still.

Also

- rewrite get_user_notifications_status for the new chats database, as the implementation is quite different.

- improved logging