Update /notify command for py-tg-bot and new db
/notify now works as fine as ever: nothing to do there. Required minimal changes to the main launchbot.py file: changes there are mainly side-effects, such as improvements and error fixes suggested by pylint.

toggle_notification required basically a complete rewrite, as the new database structure wasn't really compatible in any way. This did, however, simplify the code quite a bit.

Also, TIL: defining an immutable set (in the context of, e.g. a "if x in (a, b, c)" comparison) is way more memory inefficent that defining it as a tuple. So, changed a bunch of sets to tuples.