Add generic_update_handler
The bot now "sees" regular updates that aren't commands or text: this includes chat/group events, such as the bot being added to a group, a group being created, or a group being disbanded.

This is effectively the last missing piece from the full release of 1.6.0: the version number has thus been bumped to 1.6.0-rc2.