Edge-case bug fixes, exception handling, filters
Fixed edge-case where user attempts to send a callback via a message sent before the chat was migrated to a supergroup. This would show the sender status as "left", causing the admin/owner check to always fail. The user is now notified to send the command again.

Also adjusted the filtesr for the location and feedback handlers to only parse messages that are replies: this should reduce the excess processing we need to do quite a bit (the bot receives all regular text if it's given perms to remove messages)