Fix user time zone not being accounted for in /next
For whatever reason, the user's time zone UTC offset was shown next to the launch time, but was never accounted for. Instead, the command showed the absolute UTC time with the offset next to it, which means that an incorrect time was shown. Fixed by actually using the user's time zone to calculate the date time.