Implemented caching with redis
From now on, LaunchBot requires a running redis-server at 127.0.0.1:6379.

This commit implements caching of some high access-rate objects, such as spammers, statistics and the responses generated by /next, in an effort to reduce disk reads and the latency (and disk detoriation) it entails.

spammers.py was removed, as the spamming user IDs are now tracked in the redis db. Whether the spammer-key needs an expiration time or not remains to be seen.