Fix launches falling out of API update range
This fixes what the last two commits were attempting to fix. Launches effectively falling out of the 50 launch API request range has been an issue in the past, but this seemed to pop up again now that December is approaching.

The launches database table already stores the last time the launch row was updated, so cleaning these launches that

a) haven't launched
b) weren't updated the last time there was an API update

is as trivial as removing launches with an SQL query. In this case, this is achieved in the clean_launch_db function.

Also:

- removed the excessive logging caused by skipping scheduling for launches with TBD date+time.