Stop using tbd_date and tbd_time due to depracation
Now using statuses, as intended: tbd_date and tbd_time will be removed in a future release.

This update also updates the LL2 api from 2.0.0 to 2.1.0, which adds the TBC status: this change is reflected in the schedule command.

After the haphazard addition of holding status before the SN8 flight, this update refines the holding status and adds the flying status – should make launches much clearer.

Other changes:

- make sure redis connection exists when starting

- add cli arg to force api update

- parse cli args with argparse instead of sys.argv