Fix postpone using post-postpone notification send states
Yet another thing that's been contributing to the issues. Notifications send just fine now, but finding the chats to the notification to used the new notification states.

E.g. if notifications have been sent so that the states are "1,1,1,0" (24h=sent, 12h=sent, 60m=sent, 5m=not sent) and the launch is postponed by 24 hours, the new states would be "0,0,0,0" as expected. However, when determining if a chat had been notified of this launch already, the program would incorrectly use the new "0,0,0,0" notification states instead of the old "1,1,1,0" states.