Last fix for postpone notifications
The postpone notifications work now, but we're off by one index due to the attempted fix in an earlier commit, which still attempted to set min_enabled_index based on the new notification send states. This error seemed to cause no issues for the SpX postpone notification we sent an hour-or-so ago, so that's a good thing.

All this commit does is add more comments  and reverts min_enabled_state = enum to min_enabled_state = enum - 1, which was the correct one and was used initially.

This also renames min_enabled_state to min_recvd_notif_idx, which better describes the use case: it's an index of a state, not a state itself. Way too complicated, but let's hope the comments help a bit.

Also: version bump, comments.