Parse launches concurrently
Results in an over 300 % speed-up compared to doing it synchronously. The processing was taking unacceptably long after the natural language parser was introduced, as in over 10 seconds per one launch update. This cuts processing time down to 3 seconds, with zero fancy optimizations: just a mutex and a whole bunch of go-routines.