Fix incorrect time when scheduling post-launch checks
The time that was used to schedule the post-launch check was the time of the _next_ notification, as the current notification has already disappeared from the cache (as expected).

This commit corrects that behavior, and adds a PostLaunch{} struct type to handle them, which includes the NET of the launch we're doing a post-launch check for.

Barring this new type, there are no changes in logic.