LaunchBot 3.0.0
LaunchBot 3.0.0 is a complete rewrite of LaunchBot in Go. The main motivation behind this rewrite was improving the maintainability of the program, alongside with robustness, especially during API updates and when sending thousands of notifications.

Major new features include improved robustness of the backend, massively improved spam management, reduced susceptibility to rate-limits, improved project layout, a simplified caching system, and various other changes. A migration script is included under db/migration, which migrates a v2 database to a new v3 database.