Update chat's last active time in pre-handler
Previously, the chat's activity was updated whenever it was loaded from the cache or the disk. However, during notification sends this means that users linger in the cache for quite some time before they're cleaned. This changes the behavior so that after a send, inactive users are immediately cleaned, while active users are kept.

Logging was also reduced, so that a 500-user long cache-flush doesn't spam the logs so badly. Inactive time was also changed to an hour.