Fix extremely badly timed user-cache flush
This introduces additional checks to the conditions under which flushing the user-cache is safe. This caused a missed notification, when the user-cache was flushed immediately after a notification had just been scheduled, causing the recipient list to be null.

Also adds extra logging to the sender, and additional activity-type field to users.