Fix notification time preferences never being saved
Thanks to the feedback of a sharp user, notification time preferences are now correctly saved.

For whatever reason, Gorm takes the default value of true _over_ the current subscription value, even if it has been manually set to false. Apparently, Gorm thinks that a false is somehow equal to the value having never been set.

This effectively meant that the notification subscription values were never saved for the 24-hour, 5-minute, and postpone notifications.

The change is trivial, and performs the batch update in a single transaction.